Changed UT_sint32 and UT_uint32 to be signed and unsigned "ints" instead of "longs" (respectively).  Fixed assert for pointer list in the XAP Unix File Open / Save As dialog since it never really applied to a list of enumerated integers.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1474 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/util/xp/ut_types.h
@@ -39,8 +39,8 @@ typedef		unsigned char		UT_Byte;
 typedef		unsigned short		UT_UCSChar;	/* Unicode */
 
 typedef		unsigned short		UT_uint16;
-typedef		unsigned long		UT_uint32;
-typedef		signed long			UT_sint32;
+typedef		unsigned int		UT_uint32;
+typedef		signed int			UT_sint32;
 
 /*
 	TODO we currently use plain old C 'int' all over the place.
